much faster. see desciption
defines the whole process of DICOM conversion and renaming/copying as a function. 
then, for a subject, does this whole process for all mri files simultaneously. 
might run into issues if there are 32+ folders (runs) for a single subject, as this is the number of cores available on hipergator. though i doubt this will happen and this is easily fixed.
took 18 second for a single subject before, now takes 12 seconds for two subjects